refactor: change route name from product to products
